Use `implementation` for Gradle dependencies
While the `compile` configuration is not yet officially deprecated in
vanilla Gradle, its use is discouraged. The Gradle Android Plugin has
even gone a step further and announced it will no longer be supported
in a future release. Thus, `implementation` (which was introduced in
February 2017 with Gradle 3.4) is now used instead of `compile`.